src/server/requirements.txt
six
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.3
django>=3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
httpie==2.0.0
idna==2.8
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
django-elasticsearch-dsl>=7.0.0